refactor(cli): relocate source add/clean/content logic to _app
Move the three transport-neutral source services into src/notebooklm/_app/
(source_add.py, source_clean.py, source_content.py) and shrink the
cli/services/* modules to thin re-export wrappers, so every existing import
and patch seam keeps resolving. These were already click/rich/cli-free and
returned typed results, so this is a relocation, not a rewrite.

* source_add: input detection + URL SSRF/upload-path validation + add
  workflow. SourceAddResult.payload now builds the {"id","title","type","url"}
  source-summary dict inline (byte-identical to source_serializers.
  source_summary_payload) so the _app core stays free of the CLI adapter
  package — verified byte-stable across type codes and null fields.
* source_clean: junk-source classification + batched-deletion orchestration.
* source_content: read-only get/fulltext/guide/stale fetchers.

Patch seams preserved: source_cmd/_source_render read source_add_service.* and
source_clean_service.classify_junk_sources at call time off the wrappers, and
build_source_add_plan still takes its validate_path/looks_path_shaped callables
by injection. No cross-function monkeypatch seams exist on the moved symbols.

_app/__init__.py re-exports the three modules' public surface; CLAUDE.md repo
tree gains the _app rows and the wrapper descriptions (freshness gate green).
