+++ src/notebooklm/_app/__init__.py
@@ -47,27 +47,99 @@ from .errors import ClassifiedError, ErrorCategory, classify
 from .events import ProgressEvent, ProgressSink
 from .resolve import AmbiguousIdError, Resolution, resolve_ref, validate_id
 from .serialize import to_jsonable
+from .source_add import (
+    SourceAddExecutionPlan,
+    SourceAddFacade,
+    SourceAddPlan,
+    SourceAddResult,
+    SourceAddType,
+    SourceAddValidationError,
+    add_source,
+    build_source_add_plan,
+    execute_source_add,
+    looks_like_path,
+    validate_upload_path,
+    validate_url,
+)
+from .source_clean import (
+    CleanCandidate,
+    CleanFailure,
+    CleanStatus,
+    SourceCleanResult,
+    candidates_payload,
+    classify_junk_sources,
+    normalize_url_for_dedup,
+    run_source_clean,
+)
+from .source_content import (
+    FulltextFormat,
+    SourceFulltextPlan,
+    SourceFulltextResult,
+    SourceGetPlan,
+    SourceGetResult,
+    SourceGuidePlan,
+    SourceGuideResult,
+    SourceStalePlan,
+    SourceStaleResult,
+    execute_source_fulltext,
+    execute_source_get,
+    execute_source_guide,
+    execute_source_stale,
+)
 
 __all__ = [
     "FORMAT_EXTENSIONS",
     "AmbiguousIdError",
     "ArtifactDict",
     "ClassifiedError",
+    "CleanCandidate",
+    "CleanFailure",
+    "CleanStatus",
     "DownloadOutcome",
     "DownloadPlan",
     "DownloadPlanValidationError",
     "DownloadResult",
     "DownloadTypeSpec",
     "ErrorCategory",
+    "FulltextFormat",
     "ProgressEvent",
     "ProgressSink",
     "Resolution",
+    "SourceAddExecutionPlan",
+    "SourceAddFacade",
+    "SourceAddPlan",
+    "SourceAddResult",
+    "SourceAddType",
+    "SourceAddValidationError",
+    "SourceCleanResult",
+    "SourceFulltextPlan",
+    "SourceFulltextResult",
+    "SourceGetPlan",
+    "SourceGetResult",
+    "SourceGuidePlan",
+    "SourceGuideResult",
+    "SourceStalePlan",
+    "SourceStaleResult",
+    "add_source",
     "artifact_title_to_filename",
     "build_download_plan",
+    "build_source_add_plan",
+    "candidates_payload",
     "classify",
+    "classify_junk_sources",
     "execute_download",
+    "execute_source_add",
+    "execute_source_fulltext",
+    "execute_source_get",
+    "execute_source_guide",
+    "execute_source_stale",
+    "looks_like_path",
+    "normalize_url_for_dedup",
     "resolve_ref",
+    "run_source_clean",
     "select_artifact",
     "to_jsonable",
     "validate_id",
+    "validate_upload_path",
+    "validate_url",
 ]
